renames levelService to resetService
