+++ frontend/src/service/index.ts
@@ -2,7 +2,7 @@ import * as chatService from './chatService';
 import * as defenceService from './defenceService';
 import * as documentService from './documentService';
 import * as emailService from './emailService';
-import * as levelService from './levelService';
+import * as resetService from './resetService';
 import * as startService from './startService';
 
 export {
@@ -10,6 +10,6 @@ export {
 	defenceService,
 	documentService,
 	emailService,
-	levelService,
+	resetService,
 	startService,
 };
